+++ src/string.hpp
@@ -1,0 +1,43 @@
+/**
+ * Z80 Assembler for VGS-Zero
+ * Copyright (c) 2024, Yoji Suzuki.
+ * License under GPLv3: https://github.com/suzukiplan/vgsasm/blob/master/LICENSE.txt
+ */
+#pragma once
+#include "common.h"
+
+void trim_string(char* src)
+{
+    int i, j;
+    int len;
+    for (i = 0; ' ' == src[i]; i++);
+    if (i) {
+        for (j = 0; src[i] != '\0'; j++, i++) {
+            src[j] = src[i];
+        }
+        src[j] = '\0';
+    }
+    for (len = (int)strlen(src) - 1; 0 <= len && ' ' == src[len]; len--) {
+        src[len] = '\0';
+    }
+}
+
+std::vector<std::string> split_token(std::string str, char del)
+{
+    std::vector<std::string> elems;
+    std::string item;
+    for (char ch : str) {
+        if (ch == del) {
+            if (!item.empty()) {
+                elems.push_back(item);
+            }
+            item.clear();
+        } else {
+            item += ch;
+        }
+    }
+    if (!item.empty()) {
+        elems.push_back(item);
+    }
+    return elems;
+}
